Fix renew logic.
This PR fixes the renew logic in the systemd units. An issue can present itself
where the TLS cert expires and the renew timer tries to renew the cert but
because it is expired, it cannot do so.

The previous example of using `OnFailure=step-issue-cert@%i.service` in the
README would not work because systemd cannot start an already started unit.

We now drop the  `RemainAfterExit=true` directive on `step-issue-cert@.service`
so it will go into an inactive state after the cert is issued. This allows for
`OnFailure=step-issue-cert@%i.service` to work so I moved it into the renew
template so it works by default.

The real magic here is `ExecStartPost=systemctl try-reload-or-restart %i.service`
because this will let the `step-issue-cert@.service` unit reload or restart unit
it is creating the cert for _only_ if it is started.